refactor: useProfileStore
